pyproject.toml
[project]
name = "restai"
version = "5.0.3"
description = "RESTAI, so many 'A's and 'I's, so little time..."
authors = [{ name = "Pedro Dias", email = "petermdias@gmail.com" }]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "^1.61.0"
fastapi = "^0.115.6"
langchain = "^0.3.13"
uvicorn = "*"
python-multipart = "*"
python-dotenv = "*"
"pdfminer.six" = "*"
unstructured = "*"
pydantic = "^2.10.4"
pygpt4all = "*"
llama-cpp-python = "*"
sentence_transformers = "^2.7.0"
tiktoken = "*"
python-pptx = "*"
bs4 = "*"
selenium = "*"
yake = "^0.4.8"
passlib = "^1.7.4"
SQLAlchemy = "^2.0.35"
jq = "^1.8.0"
redis = "^5.2.1"
openpyxl = "^3.1.5"
google-cloud-aiplatform = "^1.75.0"
pymysql = "^1.1.1"
unidecode = "^1.3.8"
httpx = "0.27.0"
exceptiongroup = "^1.2.2"
wheel = "^0.42.0"
pyjwt = "^2.9.0"
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
pypdf = "^5.1.0"
ollama = "^0.3.3"
opencv-python = "^4.10.0.84"
ip-adapter = "^0.1.0"
ilock = "^1.0.3"
docx2txt = "^0.8"
llama-index-core = "^0.12.14"
llama-index-embeddings-langchain = "^0.3.0"
llama-index-vector-stores-redis = "^0.4.0"
llama-index-vector-stores-chroma = "^0.4.1"
redisvl = "^0.3.7"
chromadb = "^0.6.0"
llama-index-llms-openai = "^0.3.15"
einops = "^0.8.0"
llama-index-llms-ollama = "^0.4.2"
llama-index-postprocessor-colbert-rerank = "^0.3.0"
llama-index-llms-groq = "^0.3.1"
llama-index-embeddings-openai = "^0.3.1"
llama-index-readers-file = "^0.4.1"
deepeval = "^2.3.4"
llama-index-llms-anthropic = "^0.6.3"
llama-index-llms-litellm = "^0.3.0"
psycopg2 = "^2.9.9"
llama-index-vector-stores-pinecone = "^0.4.2"
sentry-sdk = { extras = ["fastapi"], version = "^1.45.1" }
llama-index-llms-gemini = "^0.4.2"
google-generativeai = "^0.8.3"
llama-index-llms-vertex = "^0.4.2"
llama-index-llms-azure-openai = "^0.3.0"
duckduckgo-search = "^6.2.13"
wikipedia = "^1.4.0"
llama-index-storage-chat-store-redis = "^0.4.0"
pandas = "^2.2.3"
llama-index-llms-vllm = "^0.5.0"
llama-index-multi-modal-llms-ollama = "^0.4.0"
llama-index-llms-openai-like = "^0.3.3"
paramiko = "^3.5.0"
soundfile = "^0.12.1"
llama-index-node-parser-docling = "^0.3.0"
llama-index-readers-docling = "^0.3.0"
langchain-huggingface = "^0.1.2"
llama-index-embeddings-ollama = "^0.5.0"
llama-index-readers-json = "^0.3.0"
typer = "0.12.5"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
transformers = "^4.45.1"
xformers = "^0.0.28.post2"
optimum = "1.13.1"
accelerate = "^1.0.1"
bitsandbytes = "^0.44.1"
diffusers = "0.31.0"
torch = "^2.5.0"
torchvision = "^0.20.1"
kornia = "^0.7.4"
timm = "^1.0.11"

[tool.poetry.group.dev.dependencies]
pylint = "3.0.2"
debugpy = "^1.8.6"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
